move side menu to admin layout, update story
